Create SD-JWT header in SdJwtSigner
