+++ src/main/kotlin/org/sd_jwt/KeyBasedSdJwtSigner.kt
@@ -1,5 +1,6 @@
 package org.sd_jwt
 
+import com.nimbusds.jose.JOSEObjectType
 import com.nimbusds.jose.JWSAlgorithm
 import com.nimbusds.jose.JWSHeader
 import com.nimbusds.jose.JWSSigner
@@ -9,38 +10,54 @@ import com.nimbusds.jose.crypto.RSASSASigner
 import com.nimbusds.jose.jwk.*
 
 /**
+ * Data class for setting the SD-JWT header parameters typ and cty.
+ * @param type: typ header parameter (example: JOSEObjectType("vc+sd-jwt"))
+ * @param cty:  cty header parameter (example: "credential-claims-set+json")
+ */
+data class SdJwtHeader(val type: JOSEObjectType? = null, val cty: String? = null)
+
+
+/**
  * A simple key-based implementation of an SD-JWT signer.
  *
  * @param key the private JWK for creating the JWSHeader and the JWSSigner.
+ * @param sdJwtHeader set a value for the SD-JWT header parameters 'typ' and 'cty' (optional).
  */
-class KeyBasedSdJwtSigner(private val key: JWK): SdJwtSigner {
+class KeyBasedSdJwtSigner(key: JWK, sdJwtHeader: SdJwtHeader = SdJwtHeader()): SdJwtSigner {
     private val signer: JWSSigner
-    private val header: JWSHeader
+    private val header: JWSHeader.Builder
 
     init {
         when (key.keyType) {
             KeyType.OKP -> {
                 signer = Ed25519Signer(key as OctetKeyPair)
-                header = JWSHeader.Builder(JWSAlgorithm.EdDSA).keyID(key.keyID).build()
+                header = JWSHeader.Builder(JWSAlgorithm.EdDSA).keyID(key.keyID)
             }
 
             KeyType.RSA -> {
                 signer = RSASSASigner(key as RSAKey)
-                header = JWSHeader.Builder(JWSAlgorithm.RS256).keyID(key.keyID).build()
+                header = JWSHeader.Builder(JWSAlgorithm.RS256).keyID(key.keyID)
             }
 
             KeyType.EC -> {
                 signer = ECDSASigner(key as ECKey)
-                header = JWSHeader.Builder(signer.supportedECDSAAlgorithm()).keyID(key.keyID).build()
+                header = JWSHeader.Builder(signer.supportedECDSAAlgorithm()).keyID(key.keyID)
             }
 
             else -> {
                 throw NotImplementedError("JWK signing algorithm not implemented")
             }
         }
+
+        if (sdJwtHeader.type != null) {
+            header.type(sdJwtHeader.type)
+        }
+        if (sdJwtHeader.cty != null) {
+            header.contentType(sdJwtHeader.cty)
+        }
     }
-    override fun baseHeader(): JWSHeader.Builder {
-        return JWSHeader.Builder(header)
+    override fun sdJwtHeader(): JWSHeader {
+        return header.build()
     }
 
     override fun jwsSigner(): JWSSigner {
